requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-telegram-bot==20.6
requests==2.31.0
django-environ==0.11.2
psycopg2==2.9.9